requirements.txt
numpy
python-dotenv
pandas
psycopg2
paho-mqtt